Made it a Double
Avoided unnecessary casting.

+++ DownUp/GameViewController.swift
@@ -86,7 +86,7 @@ class GameViewController: UIViewController {
             self.finish = NSDate()
             let timeTaken: Double = finish.timeIntervalSinceDate(start)
             
-            NSUserDefaults.standardUserDefaults().setFloat(Float(timeTaken), forKey: "Level2Monday")
+            NSUserDefaults.standardUserDefaults().setDouble(timeTaken, forKey: "Level2Monday")
             
             print("\(timeTaken) seconds.")
             alert("DONE")
